update: models moved outside core folder for easier future updates
