+++ backend/core/assignees/repository.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/ptracker/core"
-	"github.com/ptracker/core/models"
+	"github.com/ptracker/models"
 	"gorm.io/gorm"
 )
