+++ backend/core/tasks/repository_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/ptracker/core"
-	"github.com/ptracker/core/models"
+	"github.com/ptracker/models"
 	"github.com/ptracker/testdata"
 	"github.com/ptracker/testhelpers"
 	"github.com/ptracker/testhelpers/fixtures"
