+++ backend/testhelpers/fixtures/member.go
@@ -3,7 +3,7 @@ package fixtures
 import (
 	"fmt"
 
-	"github.com/ptracker/core/models"
+	"github.com/ptracker/models"
 )
 
 func GetMemberRow(projectID, userID, role string) models.Member {
